+++ src/basilisp/lang/compiler/nodes.py
@@ -612,7 +612,7 @@ class If(Node[SpecialForm]):
 class Import(Node[SpecialForm]):
     form: SpecialForm
     aliases: Iterable["ImportAlias"]
-    refers: Iterable["str"]
+    refers: Iterable[str]
     refer_all: bool
     env: NodeEnv = attr.field(hash=False)
     children: Sequence[kw.Keyword] = vec.EMPTY